[add] BankID integration for connect
This code will trigger a authentication process and will poll every 1 second to collect the status to confirm if the caller has authenticated yet

The Amazon Connect Block for a lambda function has a maximum timeout of 8s. This means that the lambda function needs to return with a response within 8s. If the caller hasn't authenticated by then it will return the status and order reference to be looped back in to another lambda invocation to give the caller more time to authenticate. Thes Contact Flow in Connect will determine how many attempts to give the caller before breaking out of the loop.